Add qlic_oauth.h
Add config.h for storing compile time values

Handle basic argument parsing for testing

+++ qlic_response_handler.h
@@ -4,6 +4,6 @@
 #include <stddef.h>
 
 int qlic_handle_read_chat(char* response, size_t response_size, size_t nmemb, void *userp);
-int qlic_handle_send_message(char* response, size_t response_size, size_t nmemb, void *userp);
+size_t qlic_handle_send_message(char* response, size_t response_size, size_t nmemb, void *userp);
 
 #endif
